Correct EPO WCS and FITS export
From looking at the RCW36 example, I realized that the AVM rotation
that we were emitting was negated. This wasn't obvious from the Cas A
example since its rotation is nearly zero.

After further digging, our EPO WCS wasn't correct. After some
experimentation, it appears that the right thing to do is to change the
order of the matrix multiplication to update the CD matrix. This
essentially negates the rotation operation. If we write out the EPO
image in a FITS file with exactly that WCS, we get the right results,
*if* we also stop reordering the image vertically. Since WCS can capture
both FITS and JPEG-like image parities, I think this is the correct
thing to do, since it validates that the WCS correctly applies to
exactly the EPO data as read into memory.

Once we do the above, the emitted AVM becomes correct, because it's
derived from the EPO WCS.