backend/ExportWCS.cpp: fix FITS output headers
Besides the FITS output code generally being messy, it was writing
all of the headers as strings. Some FITS viewers can deal with this,
but it's better to correctly write the numbers as numbers.

After this change, you can `toasty view` the FITS file output to preview
the alignment in WWT, before doing the full image output.