Fix rotation again?!
I feel like my brain is melting, but when I did a test with some DSS M31
imagery, I found that the rotation of the EPO WCS FITS output was
negated again. This might have surfaced because unlike the other
samples, the WCS of the input DSS M31 reference FITS image has a
non-zero rotation, while the Chandra images are all unrotated on the
sky.

The fix is to undo the CD/epo2fits matrix reordering, and negate the
rotation encoded into the AVM. This inverts the rotation of the RCW36
test case FITS file, which I was looking at in my earlier tests. I guess
I must have messed that up somehow.

With this change, the FITS files for DSS M31 and RCW31 are looking good
on the sky, as are the final AVM images. So are Cas A, which is not very
helpful since almost everything has 0 rotation, and the "Galaxy Tails"
image, which is always annoying to test since it's hard to align and
doesn't compare easily with the background imagery.